requirements.txt
pip~=23.3.1
attrs~=23.1.0
wheel~=0.38.4
cryptography~=40.0.2
Pillow~=9.5.0
tornado~=6.3.3
Jinja2~=3.1.2
Django~=4.2.2
asgiref~=3.7.2
sqlparse~=0.4.4
PyYAML~=6.0.1
numpy~=1.23.5
requests~=2.31.0
setuptools~=65.5.1
packaging~=23.1
pyparsing~=3.1.0